requirements-dev.txt
pillow==10.2.0
